refactor: ddd architecture
