+++ src/app/factory.py
@@ -2,7 +2,7 @@ from flask import Flask
 from flasgger import Swagger
 
 from src.app.schemas import ma
-from src.app.infra.logging import logging_service
+from src.app.infra.logging import logging_infra
 from src.app.configs.environ import DefaultConfig
 from src.app.utils.helpers.logging import get_logger
 from src.app.infra.middlewares.errors import register_error_handlers
@@ -14,7 +14,7 @@ log = get_logger(__name__)
 
 def create_app() -> Flask:
     app = Flask(__name__)
-    logging_service.configure_loggers()
+    logging_infra.configure_loggers()
 
     log.info("Setting app configuration")
     app.config.from_prefixed_env(prefix="")  # variables starting with an underscore will be loaded
