+++ src/app/infra/middlewares/errors.py
@@ -14,11 +14,14 @@ def handle_marshmallow_exc(e: ValidationError) -> FlaskResponse:
     first_error = list(e.messages.values())[0][0]
     return jsonify_error_response(400, "Bad Request", first_error)
 
+
 def handle_redis_exc(e: RedisError) -> FlaskResponse:
     return make_500_response("Redis exception occurred")
 
+
 def handle_any_exc(e: BaseException) -> FlaskResponse:
     return make_500_response()
+
 
 def make_500_response(msg: str = "Uncaught exception occurred"):
     log.exception(f"{msg}:")
@@ -28,6 +31,7 @@ def make_500_response(msg: str = "Uncaught exception occurred"):
         InternalServerError.description
     )
 
+
 def handle_any_error(e: HTTPException) -> FlaskResponse:
     log.error(f"Handling error: {e.__repr__()}")
     response = e.get_response()
@@ -36,6 +40,7 @@ def handle_any_error(e: HTTPException) -> FlaskResponse:
     response.data = json.dumps(data)
     response.mimetype = "application/json"
     return response
+
 
 def register_error_handlers(app: Flask) -> None:
     for ex in default_exceptions:
