+++ src/app/repositories/logging.py
@@ -1,10 +1,8 @@
 from typing import Optional
-
 from src.app.repositories.redis import RedisRepository
 
 
 class LoggingRepository(RedisRepository):
-
     def get_level(self) -> Optional[str]:
         return self.redis.db("get", "log")
