+++ src/app/schemas/logging.py
@@ -1,7 +1,7 @@
 from marshmallow.validate import OneOf
 
 from src.app.schemas import ma
-from src.app.infra.logging import logging_service
+from src.app.infra.logging import logging_infra
 from src.app.utils.helpers.logging import get_logger
 
 log = get_logger(__name__)
@@ -10,5 +10,5 @@ log = get_logger(__name__)
 class LoggingSchema(ma.Schema):
     level = ma.String(
         required=True,
-        validate=OneOf(list(logging_service.AVAILABLE_LOG_LEVELS.keys()))
+        validate=OneOf(list(logging_infra.AVAILABLE_LOG_LEVELS.keys()))
     )
