+++ src/app/services/oath/otp.py
@@ -1,10 +1,10 @@
 from time import time
 from pyotp import random_base32
-
 from src.app.services.oath import BaseOTPService
 from src.app.utils.helpers.logging import get_logger
 from src.app.infra.aes_cipher import aes_cipher_infra
-from src.app.domains.oath.otp import OTPEntity, OTPGenerator
+from src.core.entities.otp_entity import OTPEntity
+from src.core.services.otp_generator import OTPGenerator
 
 log = get_logger(__name__)
