+++ src/app/services/oath/totp.py
@@ -1,10 +1,9 @@
 from pyotp import random_base32
-
 from src.app.services.oath import BaseOTPService
 from src.app.utils.helpers.logging import get_logger
 from src.app.infra.aes_cipher import aes_cipher_infra
-from src.app.domains.oath.totp import TOTPEntity, TOTPGenerator
-
+from src.core.entities.totp_entity import TOTPEntity
+from src.core.services.totp_generator import TOTPGenerator
 
 log = get_logger(__name__)
