+++ src/app/swagger.yaml
@@ -2,7 +2,7 @@ openapi: 3.0.3
 info:
   title: open-mfa
   description: developed by github.com/rodzera
-  version: 1.3
+  version: 1.4
 servers:
   - url: /api
 paths:
