+++ src/core/services/hotp_generator.py
@@ -1,6 +1,4 @@
-from typing import Union
 from pyotp import HOTP as PyHOTP
-
 from src.app.configs.oath import OATH_CONFIG
 from src.app.utils.helpers.logging import get_logger
 
@@ -8,7 +6,6 @@ log = get_logger(__name__)
 
 
 class HOTPGenerator:
-
     def __init__(
         self,
         raw_secret: str,
@@ -46,26 +43,3 @@ class HOTPGenerator:
                 log.debug("Resync protocol successful")
                 return True, lookahead_counter
         return False, 0
-
-
-class HOTPEntity:
-    def __init__(
-        self,
-        hash_secret: str,
-        count: int,
-        resync_threshold: int
-    ):
-        self.count = int(count)
-        self.hash_secret = hash_secret
-        self.resync_threshold = int(resync_threshold)
-
-    def increment(self, amount: int = 1):
-        self.count += amount
-
-    @property
-    def as_dict(self) -> dict[str, Union[str, int]]:
-        return {
-            "secret": self.hash_secret,
-            "count": self.count,
-            "resync_threshold": self.resync_threshold
-        }
